--- front/app/postlist/page.tsx
@@ -1,9 +1,0 @@
-import ArticleList from "@/components/shared/ArticleList";
-
-export default function PostPage() {
-  return (
-    <div className="m-4">
-      <ArticleList />
-    </div>
-  );
-}
